+++ src/main/kotlin/com/github/cnrture/quickprojectwizard/arch/common/EmptyConstants.kt
@@ -1,0 +1,7 @@
+package com.github.cnrture.quickprojectwizard.arch.common
+
+fun emptyConstants(packageName: String) = """
+package $packageName.common
+
+object Constants
+""".trimIndent()