+++ src/main/kotlin/com/github/cnrture/quickprojectwizard/arch/data/source/remote/EmptyMainService.kt
@@ -1,0 +1,7 @@
+package com.github.cnrture.quickprojectwizard.arch.data.source.remote
+
+fun emptyMainService(packageName: String) = """
+package $packageName.data.source.remote
+
+interface MainService
+"""